refactor gatsby-config prismic code
